+++ src/main/java/com/standard/service/subCategoria/SubCategoriaService.java
@@ -1,19 +1,19 @@
 package com.standard.service.subCategoria;
 
-import java.util.List;
-
 import com.standard.domain.SubCategoria;
+
+import java.util.List;
 
 public interface SubCategoriaService {
 
 	SubCategoria incluir(SubCategoria objc);
 
-	SubCategoria alterar(Integer codigo, SubCategoria objc);
+	SubCategoria alterar(Long codigo, SubCategoria objc);
 
-	void excluir(Integer codigo);
+	void excluir(Long codigo);
 	
 	List<SubCategoria> consultar();
 
-	SubCategoria consultarByCodigo(Integer codigo);
+	SubCategoria consultarByCodigo(Long codigo);
 	
 }
